+++ 1/CMakeLists.txt
@@ -12,7 +12,12 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall -Wextra")
 # альтернатива: target_link_libraries(main prog1)
 #link_libraries(prog1)
 # создание исполняемого файла
-add_executable(main main.cpp)
+add_executable(main main.cpp
+        input.cpp
+        input.h
+        files.cpp
+        files.h
+        main.h)
 # для сборки из консоли:
 #
 # mkdir build # создание директории для файлов сборки
