+++ 1/files.h
@@ -1,0 +1,6 @@
+#ifndef LAB1_FILES_H
+#define LAB1_FILES_H
+
+#include "main.h"
+
+#endif //LAB1_FILES_H
